requirements.txt
streamlit==1.39.0
pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1